+++ test/DocumentFormat.OpenXml.Framework.Tests/OpenXmlNamespaceTests.cs
@@ -12,7 +12,7 @@ namespace DocumentFormat.OpenXml.Framework.Tests
         public void NamespaceCount()
         {
             var resolver = new OpenXmlNamespaceResolver();
-            Assert.Equal(151, resolver.GetNamespacesInScope(System.Xml.XmlNamespaceScope.All).Count);
+            Assert.Equal(154, resolver.GetNamespacesInScope(System.Xml.XmlNamespaceScope.All).Count);
         }
 
         [InlineData("", "", FileFormatVersions.Office2007, 0)]